requirements.txt
# Language Translation Ability Requirements

# AI/ML Dependencies
ollama>=0.1.0
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99

# HTTP Client
httpx>=0.24.0
aiohttp>=3.8.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utilities
click>=8.0.0
rich>=13.0.0
python-dotenv>=1.0.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
